Fix missing imports and move `escape-highlight-test.js` file to `utils/__tests__/`
