+++ src/lib/utils/__tests__/escape-highlight-test.js
